refactor: test config 작성 및 반영
